Change killable to destructible and kill to destroy
